+++ gui/slick/js/home.js
@@ -1,13 +1,21 @@
+/** @namespace config.sortArticle */
+/** @namespace config.homeSearchFocus */
+/** @namespace config.fuzzyDating */
+/** @namespace config.fuzzydate */
+/** @namespace config.datePreset */
+/** @namespace config.timePreset */
+/** @namespace config.isPoster */
+/** @namespace config.posterSortby */
+/** @namespace config.posterSortdir */
 $.tablesorter.addParser({
 	id: 'loadingNames',
 	is: function (s) {
 		return !1;
 	},
 	format: function (s) {
-		var name = (s || '');
-		if (0 == name.indexOf('Loading...'))
-			return name.replace('Loading...', '000');
-		return config.sortArticle ? name : name.replace(/^(?:(?:A(?!\s+to)n?)|The)\s(\w)/i, '$1');
+		var name = (s.toLowerCase() || '');
+		return (0 == name.indexOf('loading...')) ? name.replace('loading...', '000')
+			: config.sortArticle ? name : name.replace(/^(?:(?:A(?!\s+to)n?)|The)\s(\w)/i, '$1');
 	},
 	type: 'text'
 });
@@ -18,10 +26,40 @@ $.tablesorter.addParser({
 		return !1;
 	},
 	format: function (s) {
-		return s.replace('hd1080p', 5).replace('hd720p', 4).replace('hd', 3).replace('sd', 2).replace('any', 1).replace('custom', 7);
+		return s.replace('hd1080p', 22).replace('hd720p', 21).replace('hd', 20)
+			.replace('sd', 10).replace('any', 1).replace('custom', 50);
 	},
 	type: 'numeric'
 });
+
+$.tablesorter.addParser({
+	id: 'downloads',
+	is: function (s) {
+		return !1;
+	},
+	format: function(s) {
+		return valueDownloads(s);
+	},
+	type: 'numeric'
+});
+
+function valueDownloads(s) {
+	var match = s.match(/^(\?|\d+)(?:[^/]+[^\d]+(\d+))?$/);
+
+	if (null == match || '?' == match[1])
+		return -10;
+
+	var dlCnt = parseInt(match[1], 10), epsCnt = parseInt(match[2], 10);
+
+	if (0 == dlCnt)
+		return epsCnt;
+
+	var perNum = parseInt(1000000000 * parseFloat(dlCnt / epsCnt), 10), finalNum = perNum;
+	if (0 < finalNum)
+		finalNum += dlCnt;
+
+	return finalNum;
+}
 
 $(document).ready(function () {
 	if (config.homeSearchFocus) {
@@ -40,9 +78,11 @@ $(document).ready(function () {
 	}
 
 	$('div[id^="progressbar"]').each(function (k, v) {
-		var progress = parseInt($(this).siblings('span[class="sort-data"]').attr('data-progress'), 10), elId = '#' + $(this).attr('id'), v = 80;
+		var progress = parseInt($(this).siblings('span[class="sort-data"]').attr('data-progress'), 10),
+			elId = '#' + $(this).attr('id');
+		v = 80;
 		$(elId).progressbar({value: progress});
-		if (progress < 80) {
+		if (progress < v) {
 			v = progress >= 40 ? 60 : (progress >= 20 ? 40 : 20);
 		}
 		$(elId + ' > .ui-progressbar-value').addClass('progress-' + v);
@@ -60,6 +100,9 @@ $(document).ready(function () {
 					break;
 				case 'progress':
 					sortCriteria = ['progress', 'name', 'date', 'network'];
+					break;
+				case 'quality':
+					sortCriteria = ['quality', 'name', 'date', 'network', 'progress'];
 					break;
 				default:
 					sortCriteria = ['name', 'date', 'network', 'progress'];
@@ -90,8 +133,11 @@ $(document).ready(function () {
 								.replace(/^(.*?)\W*[(]\w{2,3}[)]|1$/i, '$1') || '';
 					},
 					progress: function (itemElem) {
-						var progress = $(itemElem).children('.sort-data').attr('data-progress');
-						return progress.length && parseInt(progress, 10) || Number.NEGATIVE_INFINITY;
+						var progress = $(itemElem).find('.show-dlstats').text();
+						return valueDownloads(progress);
+					},
+					quality: function (itemElem) {
+						return $(itemElem).find('.show-quality').text().toLowerCase();
 					}
 				}
 			});
@@ -123,16 +169,16 @@ $(document).ready(function () {
 				sortList: [[5, 1], [1, 0]],
 				textExtraction: {
 					0: function (node) {
-						return $(node).find('span').text().toLowerCase();
+						return $(node).find('span.sort-data').text();
 					},
 					2: function (node) {
-						return $(node).find('span').text().toLowerCase();
+						return $(node).find('span.sort-data').text().toLowerCase();
 					},
 					3: function (node) {
 						return $(node).find('span').text().toLowerCase();
 					},
 					4: function (node) {
-						return $(node).find('span').attr('data-progress');
+						return $(node).find('.progressbarText').text();
 					},
 					5: function (node) {
 						return $(node).find('i').attr('alt');
@@ -140,10 +186,9 @@ $(document).ready(function () {
 				},
 				widgets: ['saveSort', 'zebra', 'stickyHeaders', 'filter'],
 				headers: {
-					0: {sorter: 'isoDate'},
 					1: {sorter: 'loadingNames'},
 					3: {sorter: 'quality'},
-					4: {sorter: 'eps'}
+					4: {sorter: 'downloads'}
 				},
 				widgetOptions: {
 					filter_columnFilters: !1