+++ CHANGES.md
@@ -44,6 +44,7 @@
 * Update network logos to their current logo
 * Remove redundant Adult Swim logos
 * Add scene qualities WEB.h264 to SDTV, 720p.WEB.h264 to WEB DL 720p, and 1080p.WEB.h264 to WEB DL 1080p
+* Change improve handling when provider PiSexy is missing expected data
 
 
 ### 0.11.10 (2016-03-17 19:00:00 UTC)